Implement component retry properly Now we retry if we 404 or if we can't find a worker. Only one retry per error condition is allowed.
The alternative here is just to call `update_workers` on a fast
period. This has the advantage of removing a ton of code. However,
in case of a read storm, it could cause requests to erroneously
fail.

I also wonder if there is a locking strategy that would give us
more faith in the state of the worker map. This code lets it
change from under us -- which could cause unnessesary calls to
`update_workers`. However, we probably don't want to hold the lock
across a call to the worker (which could take an arbitrarly long
time), so strategies in this direction complicate the code
somewhat.